Fix setupclass for opensource tests
